feat: autoload migrations and simplify start script
no need to get the migrations from the migration files from the repo and to mount a specific volume
to the Hasura GE service. By default, when HBP starts, it checks the migrations and installs it if
required. This can be disabled by setting AUTO_MIGRATE to false. This system could be extended to
set AUTO_MIGRATE=v1 so it runs another set of migrations that would transform schema and data from
HBP v1 to HBP v2