Clean docker and migrate (#189)
* feat: autoload migrations and simplify start script

no need to get the migrations from the migration files from the repo and to mount a specific volume
to the Hasura GE service. By default, when HBP starts, it checks the migrations and installs it if
required. This can be disabled by setting AUTO_MIGRATE to false. This system could be extended to
set AUTO_MIGRATE=v1 so it runs another set of migrations that would transform schema and data from
HBP v1 to HBP v2

* fix: include mock migrations used for testing

* fix: run migrations in a separate node script

* refactor: better general repo structure, and migration that works

* ci: change test mock migration folder in GH actions