fix: correct the start command
resolveJsonModule: true is required in tsconfig.json because of the version endpoint. As
package.json is required to run HBP, tsc dist compilation ends up in creating a dist/package.json
file, but also dist/src folder. As a consequence, the start.js file path is not dist/start.js but
dist/src/start.js. The start script has been corrected in package.json, but IMO I feel this dist/src
folder is a bit ugly